updated the checkHeap method to show the byte tally statistics
refactored heap class so that the addBytes and subBytes actually updates the byte counters of the heap, added a method to clear the heap, added a method the detects if a header is the heap's header and added a method that removes a header from the linked list.

refactored the HeapManager class to have a flag that disables the manager if not in debug mode, removed unneeded comments, added overloaded addHeaderToHeap methods to add a header to the heap's linked list and update the heap's byte tally

added checks to the overloaded new and delete operators to ensure they don't operate when not in debug mode and removed unused code.