fucking finished!!! o.0 removed items and extended headers and replaced with Frames that hold a header, a footer and a pointer to the item along with linked list ponters and sizes for debug purposes.
memory manager no uses heaps and allocates a memory pool, wrapped in a frame, that is allocated on the heap.

it will only add another pool (automagically) when there is no pool of the correct size available.

clearing a heap will clear a pool recursively if any have been allocated.

Happy New Year!!!